Move menu code to separate file
